# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: di-eval-bp
host: localhost:10010
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /encrypt:
    x-swagger-router-controller: encrypt
    post:
      description: check input and resend to remote API server
      operationId: encrypt
      parameters:
        - name: plaintext
          in: body
          required: true
          schema: 
            $ref: "#/definitions/EncryptPostRequest"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EncryptPostResponse"
        "400":
          description: Bad Request, usually malformed parameters
          schema:
            $ref: "#/definitions/ErrorResponse"
        "413":
          description: Entity Too Large
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  EncryptPostRequest:
    required: 
      - plaintext
    properties:
      plaintext:
        type: string
  EncryptPostResponse:
    required:
      - ciphertext
    properties:
      ciphertext:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string